add staff sanitise to uploadFiles endpoint, move helper functions in bulk upload utils file
