+++ server/test/unit/routes/establishments/bulkUpload/uploaded.spec.js
@@ -1,192 +1,26 @@
 const expect = require('chai').expect;
 const sinon = require('sinon');
-const models = require('../../../../../models');
 const S3 = require('../../../../../routes/establishments/bulkUpload/s3');
-
-const {
-  uploadedStarGet,
-  staffData,
-  showNinoAndDob,
-  hideNinoAndDob,
-} = require('../../../../../routes/establishments/bulkUpload/uploaded');
+const buUtils = require('../../../../../utils/bulkUploadUtils');
+const { uploadedStarGet } = require('../../../../../routes/establishments/bulkUpload/uploaded');
 
 describe('/server/routes/establishment/uploaded.js', () => {
   afterEach(() => {
     sinon.restore();
   });
 
-  describe('hideNinoAndDob', () => {
-    const niNoIndex = 3;
-    const dobIndex = 5;
-
-    it('should return the data with the nino and dob set to admin, when they are present in the data', () => {
-      const dataArr = ['human', 'Nurse Jones', 'UPDATE', 'AB123456B', 'AB1 2CD', '01/02/1990'];
-      const expectedResult = ['human', 'Nurse Jones', 'UPDATE', 'Admin', 'AB1 2CD', 'Admin'];
-
-      expect(hideNinoAndDob(dataArr, niNoIndex, dobIndex)).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data with the nino and dob set to admin, when they are admin in the data', () => {
-      const dataArr = ['human', 'Nurse Jones', 'UPDATE', 'Admin', 'AB1 2CD', 'Admin'];
-      const expectedResult = ['human', 'Nurse Jones', 'UPDATE', 'Admin', 'AB1 2CD', 'Admin'];
-
-      expect(hideNinoAndDob(dataArr, niNoIndex, dobIndex)).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data with the nino and dob blank, when they are not in the data', () => {
-      const dataArr = ['human', 'Nurse Jones', 'UPDATE', '', 'AB1 2CD', ''];
-      const expectedResult = ['human', 'Nurse Jones', 'UPDATE', '', 'AB1 2CD', ''];
-
-      expect(hideNinoAndDob(dataArr, niNoIndex, dobIndex)).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data with the nino set to admin and dob blank, when the nino is the data but the dob is not', () => {
-      const dataArr = ['human', 'Nurse Jones', 'UPDATE', 'AB123456B', 'AB1 2CD', ''];
-      const expectedResult = ['human', 'Nurse Jones', 'UPDATE', 'Admin', 'AB1 2CD', ''];
-
-      expect(hideNinoAndDob(dataArr, niNoIndex, dobIndex)).to.deep.equal(expectedResult);
-    });
-  });
-
-  describe('showNinoAndDob', () => {
-    const niNoIndex = 3;
-    const dobIndex = 5;
-    const worker = { NationalInsuranceNumberValue: 'AB123456B', DateOfBirthValue: '01/02/1990' };
-
-    it('should return the data showing the nino and dob, when they are Admin in the data', () => {
-      const dataArr = ['human', 'Nurse Jones', 'UPDATE', 'Admin', 'AB1 2CD', 'Admin'];
-      const expectedResult = ['human', 'Nurse Jones', 'UPDATE', 'AB123456B', 'AB1 2CD', '01/02/1990'];
-
-      expect(showNinoAndDob(dataArr, worker, niNoIndex, dobIndex)).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data showing the nino and dob, when they are in the data', () => {
-      const dataArr = ['human', 'Nurse Jones', 'UPDATE', 'AB123456B', 'AB1 2CD', '01/02/1990'];
-      const expectedResult = ['human', 'Nurse Jones', 'UPDATE', 'AB123456B', 'AB1 2CD', '01/02/1990'];
-
-      expect(showNinoAndDob(dataArr, worker, niNoIndex, dobIndex)).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data with the nino and dob blank, when they are not in the data', () => {
-      const dataArr = ['human', 'Nurse Jones', 'UPDATE', '', 'AB1 2CD', ''];
-      const expectedResult = ['human', 'Nurse Jones', 'UPDATE', '', 'AB1 2CD', ''];
-
-      expect(showNinoAndDob(dataArr, worker, niNoIndex, dobIndex)).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data with the nino set to admin and dob blank, when the nino is the data but the dob is not', () => {
-      const dataArr = ['human', 'Nurse Jones', 'UPDATE', '', 'AB1 2CD', 'Admin'];
-      const expectedResult = ['human', 'Nurse Jones', 'UPDATE', '', 'AB1 2CD', '01/02/1990'];
-
-      expect(showNinoAndDob(dataArr, worker, niNoIndex, dobIndex)).to.deep.equal(expectedResult);
-    });
-  });
-
-  describe('staffData', () => {
-    beforeEach(() => {
-      sinon
-        .stub(models.worker, 'findOne')
-        .returns({ NationalInsuranceNumberValue: 'AB123456B', DateOfBirthValue: '01/02/1990' });
-    });
-
-    it('should return the data correctly formatted for a download type of StaffSanitise', async () => {
-      const downloadType = 'StaffSanitise';
-
-      const data = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
-
-      const expectedResult = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,Admin,AB1 2CD,Admin,`;
-
-      const result = await staffData(data, downloadType);
-
-      expect(result).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data correctly formatted for a download type of StaffSanitise when Admin in the data', async () => {
-      const downloadType = 'StaffSanitise';
-
-      const data = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,Admin,AB1 2CD,Admin,`;
-
-      const expectedResult = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,Admin,AB1 2CD,Admin,`;
-
-      const result = await staffData(data, downloadType);
-
-      expect(result).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data correctly formatted for a download type of StaffSanitise when nino and dob are not in the data', async () => {
-      const downloadType = 'StaffSanitise';
-
-      const data = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,,AB1 2CD,,`;
-
-      const expectedResult = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,,AB1 2CD,,`;
-
-      const result = await staffData(data, downloadType);
-
-      expect(result).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data correctly formatted for a download type of Staff', async () => {
-      const downloadType = 'Staff';
-
-      const data = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,Admin,AB1 2CD,Admin,`;
-
-      const expectedResult = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
-
-      const result = await staffData(data, downloadType);
-
-      expect(result).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data correctly formatted for a download type of Staff when nino and dob in the data', async () => {
-      const downloadType = 'Staff';
-
-      const data = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
-
-      const expectedResult = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
-
-      const result = await staffData(data, downloadType);
-
-      expect(result).to.deep.equal(expectedResult);
-    });
-
-    it('should return the data correctly formatted for a download type of Staff when nino and dob are not in the data', async () => {
-      const downloadType = 'Staff';
-
-      const data = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,,AB1 2CD,,`;
-
-      const expectedResult = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,,AB1 2CD,,`;
-
-      const result = await staffData(data, downloadType);
-
-      expect(result).to.deep.equal(expectedResult);
-    });
-  });
-
   describe.only('uploadedStarGet', () => {
-    beforeEach(() => {
-      sinon
-        .stub(models.worker, 'findOne')
-        .returns({ NationalInsuranceNumberValue: 'AB123456B', DateOfBirthValue: '01/02/1990' });
-    });
-
-    describe('downloadType = StaffSanitise', () => {
+    describe.only('downloadType = StaffSanitise', () => {
       it('returns sanitised staff file when dob and nino are in data', async () => {
         const data = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
         human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
 
         sinon.stub(S3, 'downloadContent').returns({ data });
+
+        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
+        human,Nurse Jones,UPDATE,Admin,AB1 2CD,Admin,`;
+
+        sinon.stub(buUtils, 'staffData').returns(updatedData);
 
         const saveResponse = sinon.stub(S3, 'saveResponse');
 
@@ -197,9 +31,6 @@ describe('/server/routes/establishment/uploaded.js', () => {
           },
           {},
         );
-
-        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,Admin,AB1 2CD,Admin,`;
 
         expect(saveResponse.getCalls()[0].args).to.deep.equal([
           {
@@ -222,6 +53,11 @@ describe('/server/routes/establishment/uploaded.js', () => {
 
         sinon.stub(S3, 'downloadContent').returns({ data });
 
+        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
+        human,Nurse Jones,UPDATE,Admin,AB1 2CD,Admin,`;
+
+        sinon.stub(buUtils, 'staffData').returns(updatedData);
+
         const saveResponse = sinon.stub(S3, 'saveResponse');
 
         await uploadedStarGet(
@@ -231,9 +67,6 @@ describe('/server/routes/establishment/uploaded.js', () => {
           },
           {},
         );
-
-        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,Admin,AB1 2CD,Admin,`;
 
         expect(saveResponse.getCalls()[0].args).to.deep.equal([
           {
@@ -256,6 +89,11 @@ describe('/server/routes/establishment/uploaded.js', () => {
 
         sinon.stub(S3, 'downloadContent').returns({ data });
 
+        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
+        human,Nurse Jones,UPDATE,,AB1 2CD,,`;
+
+        sinon.stub(buUtils, 'staffData').returns(updatedData);
+
         const saveResponse = sinon.stub(S3, 'saveResponse');
 
         await uploadedStarGet(
@@ -265,9 +103,6 @@ describe('/server/routes/establishment/uploaded.js', () => {
           },
           {},
         );
-
-        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,,AB1 2CD,,`;
 
         expect(saveResponse.getCalls()[0].args).to.deep.equal([
           {
@@ -285,12 +120,17 @@ describe('/server/routes/establishment/uploaded.js', () => {
       });
     });
 
-    describe('downloadType = Staff', () => {
+    describe.only('downloadType = Staff', () => {
       it('returns staff file when dob and nino are in data', async () => {
         const data = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
         human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
 
         sinon.stub(S3, 'downloadContent').returns({ data });
+
+        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
+        human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
+
+        sinon.stub(buUtils, 'staffData').returns(updatedData);
 
         const saveResponse = sinon.stub(S3, 'saveResponse');
 
@@ -301,9 +141,6 @@ describe('/server/routes/establishment/uploaded.js', () => {
           },
           {},
         );
-
-        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
 
         expect(saveResponse.getCalls()[0].args).to.deep.equal([
           {
@@ -326,6 +163,11 @@ describe('/server/routes/establishment/uploaded.js', () => {
 
         sinon.stub(S3, 'downloadContent').returns({ data });
 
+        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
+        human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
+
+        sinon.stub(buUtils, 'staffData').returns(updatedData);
+
         const saveResponse = sinon.stub(S3, 'saveResponse');
 
         await uploadedStarGet(
@@ -336,13 +178,10 @@ describe('/server/routes/establishment/uploaded.js', () => {
           {},
         );
 
-        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,AB123456B,AB1 2CD,01/02/1990,`;
-
         expect(saveResponse.getCalls()[0].args).to.deep.equal([
           {
             params: ['1/lastBulkUpload/2022-01-01-sfc-bu-staff.csv'],
-            query: { downloadType: 'Staff' },
+            query: { bulkUploadUtils: 'Staff' },
           },
           {},
           200,
@@ -360,6 +199,11 @@ describe('/server/routes/establishment/uploaded.js', () => {
 
         sinon.stub(S3, 'downloadContent').returns({ data });
 
+        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
+        human,Nurse Jones,UPDATE,,AB1 2CD,,`;
+
+        sinon.stub(buUtils, 'staffData').returns(updatedData);
+
         const saveResponse = sinon.stub(S3, 'saveResponse');
 
         await uploadedStarGet(
@@ -369,9 +213,6 @@ describe('/server/routes/establishment/uploaded.js', () => {
           },
           {},
         );
-
-        const updatedData = `LOCALESTID,UNIQUEWORKERID,STATUS,NINUMBER,POSTCODE,DOB\r\n
-        human,Nurse Jones,UPDATE,,AB1 2CD,,`;
 
         expect(saveResponse.getCalls()[0].args).to.deep.equal([
           {
@@ -389,7 +230,7 @@ describe('/server/routes/establishment/uploaded.js', () => {
       });
     });
 
-    describe.only('downloadType = Workplace', () => {
+    describe('downloadType = Workplace', () => {
       it('returns establishment file', async () => {
         const data = `LOCALESTID,STATUS,ESTNAME,ADDRESS1,ADDRESS2,ADDRESS3,POSTTOWN,POSTCODE,\r\n
         human,UPDATE,Care Home 1,31 Some Street,,,Leeds,LS1 2AD,`;
