+++ src/app/core/services/bulk-upload.service.ts
@@ -148,6 +148,12 @@ export class BulkUploadService {
   public getUploadedFileFromS3(workplaceUid: string, key: string, type: BulkUploadFileType): Observable<any> {
     const params = new HttpParams().set('downloadType', type);
 
+    console.log('getUploadedFileFromS3 **************');
+    console.log(workplaceUid);
+    console.log(key);
+    console.log(type);
+    console.log(this.endpoint);
+
     return this.checkLockStatus(
       () => this.http.get<Blob>(`/api/establishment/${workplaceUid}/bulkupload/${this.endpoint}/${key}`, { params }),
       {
