+++ src/app/features/bulk-upload/drag-and-drop-files-list/drag-and-drop-files-list.component.ts
@@ -213,6 +213,7 @@ export class DragAndDropFilesListComponent implements OnInit, OnDestroy {
       .getUploadedFileFromS3(this.establishmentService.primaryWorkplace.uid, key, type)
       .pipe(take(1))
       .subscribe((response) => {
+        console.log(response);
         const filename = FileUtil.getFileName(response);
         const blob = new Blob([response.body], { type: 'text/plain;charset=utf-8' });
         saveAs(blob, filename);
@@ -247,6 +248,7 @@ export class DragAndDropFilesListComponent implements OnInit, OnDestroy {
         }
 
         this.uploadedFiles = uploadedFiles;
+        console.log(this.uploadedFiles);
         this.totalErrors = this.uploadedFiles.map((f) => f.errors).reduce((p, c) => c + p);
 
         if (['PASSED', 'WARNINGS', 'FAILED'].includes(state)) {
