+++ src/app/features/bulk-upload/last-bulk-upload/last-bulk-upload.component.ts
@@ -58,6 +58,7 @@ export class LastBulkUploadComponent implements OnInit {
       .getUploadedFileFromS3(this.establishmentService.primaryWorkplace.uid, key, type)
       .pipe(take(1))
       .subscribe((response) => {
+        console.log(response);
         const filename = FileUtil.getFileName(response);
         const blob = new Blob([response.body], { type: 'text/plain;charset=utf-8' });
         saveAs(blob, filename);
